refactor: move image related dtos to common
- 기존 post 이미지 관련된 dto가 ai에서도 재사용되어 common 패키지로 옮김
