+++ src/main/java/hanium/modic/backend/web/common/image/dto/response/CreateImageSaveUrlResponse.java
@@ -1,4 +1,4 @@
-package hanium.modic.backend.web.post.dto.response;
+package hanium.modic.backend.web.common.image.dto.response;
 
 public record CreateImageSaveUrlResponse(
 	String imageSaveUrl,
